Update Login.html
Removed the nav bar so there are no links to the other pages, this prevents users who haven't logged in from accessing the main pages. Also changed the hidden hyperlink from the logo so when clicked it returns to the login page not the main page to prevent unauthorised users accessing the asset maps.